requirements.txt
discord.py==1.7.2
dnspython==2.1.0
PyNaCl
async-timeout==3.0.1
python-dotenv~=0.17.0
apscheduler~=3.7.0
urbandictionary
wikipedia
python-dateutil
nest-asyncio
youtube_dl
pytz~=2021.1